+++ org.gitools.heatmap/src/main/java/org/gitools/heatmap/header/HeatmapDecoratorHeader.java
@@ -30,6 +30,8 @@ import org.gitools.heatmap.decorator.DetailsDecoration;
 import org.gitools.heatmap.decorator.impl.LinearDecorator;
 import org.gitools.matrix.filter.AnnotationFunction;
 import org.gitools.utils.formatter.HeatmapTextFormatter;
+import org.gitools.utils.formatter.ITextFormatter;
+import org.gitools.utils.formatter.ScientificHeatmapTextFormatter;
 
 import javax.xml.bind.annotation.*;
 import java.awt.*;
@@ -41,6 +43,7 @@ import java.util.List;
 @XmlRootElement
 public class HeatmapDecoratorHeader extends HeatmapHeader {
 
+    public static final String PROPERTY_NUMERIC_FORMATTER = "numericFormatter";
     public enum LabelPositionEnum {
         leftOf,
         rightOf,
@@ -58,6 +61,9 @@ public class HeatmapDecoratorHeader extends HeatmapHeader {
 
     @XmlElement
     private List<String> annotationLabels;
+
+
+    private transient ITextFormatter numericFormatter = ScientificHeatmapTextFormatter.INSTANCE;
 
     @XmlTransient
     private String sortLabel;
@@ -121,7 +127,7 @@ public class HeatmapDecoratorHeader extends HeatmapHeader {
 
         decorator.decorate(
                 decoration,
-                HeatmapTextFormatter.TWO_DECIMALS,
+                numericFormatter,
                 annotations,
                 annotations.getLayers().get(annotation),
                 identifier);
@@ -201,6 +207,20 @@ public class HeatmapDecoratorHeader extends HeatmapHeader {
         }
 
     }
+
+
+    public ITextFormatter getNumericFormatter() {
+        if (numericFormatter == null) {
+            return ScientificHeatmapTextFormatter.INSTANCE;
+        }
+        return numericFormatter;
+    }
+
+    public void setNumericFormatter(ITextFormatter numericFormatter) {
+        this.numericFormatter = numericFormatter;
+        firePropertyChange(PROPERTY_NUMERIC_FORMATTER, null, numericFormatter);
+    }
+
 
     @Override
     public boolean isNumeric() {